[Chore] Change folder structure for separate input and render from main game engine folder
