+++ platform/linux/include/render/linux_render.hpp
