+++ platform/linux/source/input/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(input linux_input.cpp)
+
+target_include_directories(input
+    PUBLIC ${PROJECT_SOURCE_DIR}/include
+    PUBLIC ${PROJECT_SOURCE_DIR}/platform/linux/include
+)
