+++ platform/linux/source/render/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(render linux_render.cpp)
+
+target_include_directories(render
+    PUBLIC ${PROJECT_SOURCE_DIR}/include
+    PUBLIC ${PROJECT_SOURCE_DIR}/platform/linux/include
+)
