+++ platform/linux/source/render/linux_render.cpp
