--- source/input/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_library(input linux_input.cpp)
-
-target_include_directories(input PUBLIC ${PROJECT_SOURCE_DIR}/include)
