--- source/render/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_library(render linux_render.cpp)
-
-target_include_directories(render PUBLIC ${PROJECT_SOURCE_DIR}/include)
