+++ source/tetris/CMakeLists.txt
@@ -5,3 +5,7 @@ add_executable(tetris
 target_link_libraries(tetris
 	PRIVATE engine
 )
+
+if (UNIX)
+    target_include_directories(tetris PUBLIC ${PROJECT_SOURCE_DIR}/platform/linux/include)
+endif()
